+++ src/main/java/com/avengers/gamera/dto/article/ArticleGetDto.java
@@ -31,5 +31,5 @@ public class ArticleGetDto {
     private EArticleType type;
     private OffsetDateTime createdTime;
     private OffsetDateTime updatedTime;
-    private Long likeNum;
+    private int likeNum;
 }
